\name{cgdefault}
\alias{cgdefault}
\encoding{UTF-8}
\title{set control defaults}
\concept{minimization}
\concept{maximization}
\description{
  Set control defaults.
}
\usage{
   cgdefault(npar)
}
\arguments{
 \item{npar}{Number of parameters to optimize.}
}
\value{
  \code{ctrldefault} returns the default control settings for optimization tools.
}
\keyword{nonlinear}
\keyword{optimize}

